Clean up code
* Update deps
* Comment out old Rocket code
* Restructure code into modules
