+++ .env.example
@@ -1,2 +1,7 @@
-DATABASE_URL=postgres://user:password@localhost/database
 AMQP_ADDRESS=127.0.0.1:5672
+
+POSTGRES_USER=import
+POSTGRES_PASSWORD=BSStykoth4XneRoAguZx
+POSTGRES_DB=schani_import
+
+DATABASE_URL=postgres://import:BSStykoth4XneRoAguZx@localhost:54322/schani_import
