+++ .env.testing
@@ -1,2 +1,7 @@
-DATABASE_URL=postgres://travis@localhost/schani_import
 AMQP_ADDRESS=127.0.0.1:5672
+
+POSTGRES_USER=import
+POSTGRES_PASSWORD=BSStykoth4XneRoAguZx
+POSTGRES_DB=schani_import
+
+DATABASE_URL=postgres://import:BSStykoth4XneRoAguZx@localhost:54322/schani_import
