+++ Cargo.toml
@@ -1,25 +1,21 @@
 [package]
 name = "schani_import"
-version = "0.1.0"
+version = "0.2.0"
 authors = ["Christoph Wurst <christoph@winzerhof-wurst.at>"]
 
 [dependencies]
-amq-protocol = "^0.16"
-diesel = { version = "0.12.0", features = ["postgres"] }
-diesel_codegen = { version = "0.11.0", features = ["postgres"] }
-dotenv = "0.8.0"
+amq-protocol = "^0.19"
+diesel = { version = "1.0.0-beta1", features = ["postgres"] }
+dotenv = "0.10.1"
 futures = "^0.1"
-hyper = "^0.10"
-lapin-futures = "^0.8"
-log = "^0.3"
+hyper = "^0.11"
+lapin-futures = "^0.10"
+log = "*"
 resolve = "0.1.2"
-rocket = "0.2.8"
-rocket_codegen = "0.2.8"
-rocket_contrib = "0.2.6"
 serde = "0.9"
-serde_derive = "0.9"
-serde_json="1.0.2"
-serde_urlencoded = "0.5.1"
-tokio-core = "0.1.7"
+serde_derive = "^1.0"
+serde_json="^1.0"
+serde_urlencoded = "^0.5"
+tokio-core = "^0.1"
 tokio-io = "^0.1"
 tokio-timer = "^0.1"
