--- src/schema.rs
@@ -1,1 +1,0 @@
-infer_schema!("dotenv:DATABASE_URL");
