+++ src/service/store.rs
@@ -1,8 +1,8 @@
-use super::models::Import;
-
 use std::error::Error;
 use std::io::Read;
 use std::io::Write;
+
+use models::Import;
 
 use hyper;
 use hyper::client::Client;
@@ -16,14 +16,12 @@ fn build_new_raw_image_body(
     longitude: f64,
     latitude: f64,
 ) -> String {
-    serde_urlencoded::to_string(
-        [
-            ("user_id", user_id.to_string()),
-            ("camera", camera.to_owned()),
-            ("longitude", longitude.to_string()),
-            ("latitude", latitude.to_string()),
-        ],
-    ).unwrap()
+    serde_urlencoded::to_string([
+        ("user_id", user_id.to_string()),
+        ("camera", camera.to_owned()),
+        ("longitude", longitude.to_string()),
+        ("latitude", latitude.to_string()),
+    ]).unwrap()
 }
 
 fn save_raw_image_data(import: &Import) -> Result<i32, String> {
@@ -49,12 +47,14 @@ fn save_raw_image_data(import: &Import) -> Result<i32, String> {
     }
 
     let mut resp_text = String::new();
-    try!(resp.read_to_string(&mut resp_text).map_err(|err| {
-        format!("could not read response: {}", err)
-    }));
-    let resp_json: Value = try!(serde_json::from_str(&resp_text).map_err(|err| {
-        format!("could not read response JSON: {}", err)
-    }));
+    try!(
+        resp.read_to_string(&mut resp_text)
+            .map_err(|err| { format!("could not read response: {}", err) })
+    );
+    let resp_json: Value = try!(
+        serde_json::from_str(&resp_text)
+            .map_err(|err| { format!("could not read response JSON: {}", err) })
+    );
     let id = resp_json["id"].as_u64().unwrap() as i32;
 
     Ok(id)
@@ -67,15 +67,13 @@ fn build_new_image_body(
     side_car_file: String,
     raw_image_id: i32,
 ) -> String {
-    serde_urlencoded::to_string(
-        [
-            ("title", title.to_owned()),
-            ("description", description.to_owned()),
-            ("license", license.to_owned()),
-            ("side_car_file", side_car_file.to_owned()),
-            ("raw_image_id", raw_image_id.to_string()),
-        ],
-    ).unwrap()
+    serde_urlencoded::to_string([
+        ("title", title.to_owned()),
+        ("description", description.to_owned()),
+        ("license", license.to_owned()),
+        ("side_car_file", side_car_file.to_owned()),
+        ("raw_image_id", raw_image_id.to_string()),
+    ]).unwrap()
 }
 
 fn save_image_data(
@@ -108,12 +106,14 @@ fn save_image_data(
     }
 
     let mut resp_text = String::new();
-    try!(resp.read_to_string(&mut resp_text).map_err(|e| {
-        format!("{}", e.description())
-    }));
-    let resp_json: Value = try!(serde_json::from_str(&resp_text).map_err(|err| {
-        format!("could not read response JSON: {}", err)
-    }));
+    try!(
+        resp.read_to_string(&mut resp_text)
+            .map_err(|e| { format!("{}", e.description()) })
+    );
+    let resp_json: Value = try!(
+        serde_json::from_str(&resp_text)
+            .map_err(|err| { format!("could not read response JSON: {}", err) })
+    );
     let id = resp_json["id"].as_u64().unwrap() as i32;
 
     Ok(id)
@@ -123,13 +123,13 @@ fn save_raw_file(raw_file_id: i32, data: &mut Read) -> Result<(), Box<Error>> {
     let mut buffer = vec![];
     try!(data.read_to_end(&mut buffer));
 
-    let image_url = try!(hyper::Url::parse(
-        &format!("http://store:8000/api/raw_images/{}", raw_file_id),
-    ));
+    let image_url = try!(hyper::Url::parse(&format!(
+        "http://store:8000/api/raw_images/{}",
+        raw_file_id
+    ),));
     let mut req = try!(hyper::client::Request::new(Method::Post, image_url));
-    req.headers_mut().set(hyper::header::ContentLength(
-        buffer.len() as u64,
-    ));
+    req.headers_mut()
+        .set(hyper::header::ContentLength(buffer.len() as u64));
     let mut stream = try!(req.start());
     try!(stream.write_all(buffer.as_slice()));
     try!(stream.flush());
@@ -149,9 +149,8 @@ fn save_sidecar_file(image_id: i32, data: &mut Read) -> Result<(), Box<Error>> {
         image_id
     )));
     let mut req = try!(hyper::client::Request::new(Method::Post, image_url));
-    req.headers_mut().set(hyper::header::ContentLength(
-        buffer.len() as u64,
-    ));
+    req.headers_mut()
+        .set(hyper::header::ContentLength(buffer.len() as u64));
     let mut stream = try!(req.start());
     try!(stream.write_all(buffer.as_slice()));
     try!(stream.flush());
