+++ src/web/extractors.rs
@@ -1,0 +1,1 @@
+s
